Fix warnings.

+++ kstars/htmesh/MeshBuffer.cpp
@@ -34,7 +34,7 @@ int MeshBuffer::append(Trixel trixel)
 }
 
 void MeshBuffer::fill() {
-    for (Trixel i = 0; i < maxSize; i++) {
+    for (Trixel i = 0; i < (unsigned int)maxSize; i++) {
         m_buffer[i] = i;
     }
     m_size = maxSize;
